Change to brainmask erosion
Erodes the brainmask twice.  Also dilates the brainmask 6 times, then erodes it 12 times.  Then generates a final mask as the intersection (logical AND) between the twice eroded brainmask and 6-dilated/12-eroded brainmask.  The reason was simply to erode substantially in cortex but not in the region that the striatum should be, and the substantial dilation helped to fill in the region between temporal cortices.